+++ server.go
@@ -1,7 +1,6 @@
 package sseserver
 
 import (
-	"log"
 	"net/http"
 )
 
@@ -53,24 +52,6 @@ func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	mux.ServeHTTP(w, r)
 }
 
-// Serve is a convenience method to begin serving connections on specified address.
-//
-// This method blocks forever, as it is basically a convenience wrapper around
-// http.ListenAndServe(addr, self).
-//
-// It also implements basic request logging to STDOUT.
-//
-// If you want to do something more sophisticated, you should not use this method,
-// but rather just build your own HTTP routing/middleware chain around Server which
-// implements the standard http.Handler interface.
-func (s *Server) Serve(addr string) {
-	log.Println("Starting server on addr " + addr)
-	handler := ProxyRemoteAddrHandler(requestLogger(s))
-	if err := http.ListenAndServe(addr, handler); err != nil {
-		log.Fatal("ListenAndServe:", err)
-	}
-}
-
 // CONSIDER: func (s *Server) Drain() for allowing active connections to drain. This
 // will require some thinking since SSE connections can be *very* long lived.
 // Handling this via k8s may work by default looking at active HTTP connections.
@@ -81,41 +62,4 @@ func (s *Server) Serve(addr string) {
 // closed in the background.
 func (s *Server) Shutdown() {
 	s.hub.Shutdown()
-}
-
-// ProxyRemoteAddrHandler is HTTP middleware to determine the actual RemoteAddr
-// of a http.Request when your server sits behind a proxy or load balancer.
-//
-// When utilized, the value of RemoteAddr will be overridden based on the
-// X-Real-IP or X-Forwarded-For HTTP header, which can be a comma separated list
-// of IPs.
-//
-// See http://httpd.apache.org/docs/2.2/mod/mod_proxy.html#x-headers for
-// details.
-//
-// Based on http://git.io/xDD3Mw
-func ProxyRemoteAddrHandler(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ip := r.Header.Get("X-Real-IP")
-		if ip == "" {
-			ip = r.Header.Get("X-Forwarded-For")
-		}
-		if ip != "" {
-			r.RemoteAddr = ip
-		}
-		next.ServeHTTP(w, r)
-	})
-}
-
-// requestLogger is a sample of integrating logging via HTTP middleware.
-//
-// Utilized in our Serve() convenience function. Note that due to the long
-// connection time of SSE requests you likely want to log disconnection
-// separately.
-func requestLogger(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		log.Println("CONNECT\t", r.URL.Path, "\t", r.RemoteAddr)
-		next.ServeHTTP(w, r)
-		log.Println("DISCONNECT\t", r.URL.Path, "\t", r.RemoteAddr)
-	})
 }
